Initial commit with just the README
Starting out with some readme driven development to clear up my head
on how this thing has to behave.

I want to try some stuff like the faraday lib and the way it is
basically built. This will probably end up being a very ruby developer
specific tool rather than a general tool, but this is just for fun, so
lets do it weird.