Commit ranges are calculated correctly now
The intent here is to commit as less as possible, so, If I have 1 commit
and my goal is to be on the :low range, I do not need to commit at all.

I needed to add a class to handle the range pair so I can just check if
we are over or under the minimum required.

There may be some fixes to do when applying this to a real repo, but so
far this looks ok from the intent point of view, I will review later the
real need for commits when it is actually being done